[feature] add nix flake for building package
Changes: Ignore git dependency only in the special case that the program
is found in a nix builder! This is done because the nix builder drops
the .git subdirectory for "purity" reasons.

How do we fetch subvolumes then?
We do that by specifying ?submodules=1 on the end of the flakeref (flake
reference) in our nix command.

Previously, we relied on a compiler pointer hack to get the short PATH
of the filenames inside that we show in the log messages. This is impure
under nix since it reflects the build environment, instead we give GCC a
flag which we use to tell the compiler what the relative path is, thus
avoiding this issue entirely.